+++ build-std.sh
@@ -10,10 +10,15 @@ else
     TARGETDIR=$1
 fi
 
+_INPLACE=0
 if [ -z "$2" ] ; then
     BRANCH="master"
 else
-    BRANCH=$2
+    if [ "xinplace" == "x$2" ] ; then
+        _INPLACE=1
+    else
+        BRANCH=$2
+    fi
 fi
 
 if [ -z "$3" ] ; then
@@ -29,28 +34,36 @@ else
 fi
 
 
-if [ -d "pmix-standard" ] ; then
+if [[ $_INPLACE == 0 && -d "pmix-standard" ]] ; then
     echo "--------- Remove old directory"
     rm -rf pmix-standard
 fi
 
 
-if [[ "$REPO" = /* ]] ; then
-    echo "--------- Copy working directory ($BRANCH)"
-    cp -R $REPO pmix-standard
+if [ $_INPLACE == 0 ] ; then
+    if [[ "$REPO" = /* ]] ; then
+        echo "--------- Copy working directory ($BRANCH)"
+        cp -R $REPO pmix-standard
+    else
+        echo "--------- Clone the branch ($BRANCH)"
+        echo git clone -b $BRANCH $REPO
+        git clone -b $BRANCH $REPO
+    fi
 else
-    echo "--------- Clone the branch ($BRANCH)"
-    echo git clone -b $BRANCH $REPO
-    git clone -b $BRANCH $REPO
+    echo "--------- Building inplace"
 fi
 
 echo "--------- Build"
-cd pmix-standard
+if [ $_INPLACE == 0 ] ; then
+    cd pmix-standard
+else
+    cd $TARGETDIR
+fi
 make clean
 make
 
 
-if [ -d "$TARGETDIR" ] ; then
+if [[ $_INPLACE == 0 && -d "$TARGETDIR" ]] ; then
     echo "--------- Copy out the pdf"
     cp pmix-standard.pdf ${TARGETDIR}/${OUT_FNAME}
 fi
